requirements.txt
# MistSiteDashboard Dependencies
# Core API and networking
mistapi>=0.3.0
requests>=2.28.0

# Web framework
Flask>=2.3.0
Werkzeug>=2.3.0

# Environment configuration
python-dotenv>=1.0.0

# Secure credential storage (required by mistapi)
keyring>=25.0.0
